backend/pyproject.toml
[project]
name = "ai-quiz-generator"
version = "0.1.0"
description = "Production-ready backend-only AI Quiz Generator (FastAPI + Groq + RAG)"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "groq>=0.9.0",
  "sentence-transformers>=3.0.0",
  "numpy<2",
  "faiss-cpu>=1.8.0.post1",
  "tiktoken>=0.7.0",
  "pypdf>=4.2.0",
  "python-docx>=1.1.2",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "tenacity>=8.3.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["static*", "storage*", "docs*", "tests*"]

[tool.uvicorn]
factory = false

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
